requirements.txt
phidata
phi
python-dotenv
packaging
duckduckgo-search
fastapi
PyMuPDF
uvicorn
groq
together
youtube-search-python
python-multipart
jinja2 python-multipart
markdown
youtube_transcript_api

# pip uninstall youtube-search-python httpx ( if yputube search doesnt work)
#>> pip install youtube-search-python==1.6.6 httpx==0.23.0